+++ src/isla/parser.py
@@ -1,5 +1,6 @@
 import itertools as I
 import random
+from functools import lru_cache
 from typing import Tuple, Iterable, Generator, List, Dict, Collection
 
 from isla.helpers import tree_to_string, RE_NONTERMINAL
@@ -77,7 +78,7 @@ class Parser:
            To be defined in subclasses."""
         raise NotImplementedError()
 
-    def parse(self, text: str) -> Iterable[ParseTree]:
+    def parse(self, text: str) -> List[ParseTree]:
         """Parse `text` using the grammar.
            Return an iterable of parse trees."""
         cursor, forest = self.parse_prefix(text)
@@ -445,3 +446,33 @@ class EnhancedExtractor(SimpleExtractor):
             if parse_tree is not None:
                 return self.parser.prune_tree(parse_tree)
         return None
+
+
+class PEGParser(Parser):
+    def parse_prefix(self, text):
+        cursor, tree = self.unify_key(self.start_symbol(), text, 0)
+        return cursor, [tree]
+
+    def unify_rule(self, rule, text, at):
+        if self.log:
+            print('unify_rule: %s with %s' % (repr(rule), repr(text[at:])))
+        results = []
+        for token in rule:
+            at, res = self.unify_key(token, text, at)
+            if res is None:
+                return at, None
+            results.append(res)
+        return at, results
+
+    @lru_cache(maxsize=None)
+    def unify_key(self, key, text, at=0):
+        if key not in self.cgrammar:
+            if text[at:].startswith(key):
+                return at + len(key), (key, [])
+            else:
+                return at, None
+        for rule in self.cgrammar[key]:
+            to, res = self.unify_rule(rule, text, at)
+            if res is not None:
+                return (to, (key, res))
+        return 0, None